rating and admin ui updated

+++ controller/slotController.js
@@ -10,10 +10,10 @@ export const addSlot= async (req,res,next)=>{
 	const {time,date,price,duration} = req.body
 
 	if(!time){
-		return res.status(403).json("add time")
+		return res.status(403).json({msg:"add time"})
 	}
 	if(!date){
-		return res.status(403).json("add date")
+		return res.status(403).json({msg:"add date"})
 	}
 
 	try{
@@ -40,9 +40,15 @@ export const getSlots= async (req,res,next)=>{
 export const deleteSlot= async (req,res,next)=>{
 	const {id} = req.user
 	const {ID} = req.params
+	if(!id){
+		return res.status(401).json({msg:"Login Again"})
+	}
+	if(!ID){
+		return res.status(401).json({msg:"This Slot doesn't exist"})
+	}
 	try{
 		 const slots = await Slot.deleteOne({_id:ID,by:id})
-	     return res.status(200).json(slots);
+	     return res.status(200).json({msg:"Slot Deleted Successfully"});
 	}catch(err){
 		next(err)
 	}
@@ -55,17 +61,21 @@ export const scheduleSlot= async (req,res,next)=>{
 
 	try{
 		 if(!id){
-		 return res.status(403).json("login first");
+		 return res.status(401).json({msg:"login first"});
          }
          if(!ID){
-		 return res.status(403).json("Invalid slot");
+		 return res.status(403).json({msg:"Invalid slot"});
          }
          if(id===by){
-		    return res.status(403).json("you can't interview your self");
+		    return res.status(403).json({msg:"you can't interview your self"});
          }
 
-	 	 const updateSlot = await Slot.findOneAndUpdate({_id:ID},{booked:booked,bookedBy:id,topic:topic},{new:true})
-	     return res.status(200).json(updateSlot);
+         if(booked){
+         	return  res.status(403).json({msg:"Already Booked"});
+         }
+
+	 	 const updateSlot = await Slot.findOneAndUpdate({_id:ID},{booked:true,bookedBy:id,topic:topic},{new:true})
+	     return res.status(200).json({msg:"Slot Booked"});
 	}catch(err){
 		next(err)
 	}
@@ -76,7 +86,7 @@ export const activeSlots= async (req,res,next)=>{
 
 	try{
 		 if(!id){
-		 return res.status(200).json("login first");
+		 return res.status(200).json({msg:"login first"});
          }
 	 	 const updateSlot = await Slot.find({by:id,booked:true}).populate({path:'bookedBy',select:'username email roles'})
 	     return res.status(200).json(updateSlot);
@@ -92,10 +102,10 @@ export const approveSlot= async (req,res,next)=>{
 
 	try{
 		 if(!id){
-		 return res.status(200).json("login first");
+		 return res.status(200).json({msg:"login first"});
          }
          if(!ID){
-		 return res.status(200).json("Invalid slot");
+		 return res.status(200).json({msg:"Invalid slot"});
          }
 	 	 const updateSlot = await Slot.findOneAndUpdate({_id:ID},{approved:approved},{new:true})
 	     return res.status(200).json(updateSlot);
@@ -109,7 +119,7 @@ export const approvedSlots= async (req,res,next)=>{
 
 	try{
 		 if(!id){
-		 return res.status(200).json("login first");
+		 return res.status(200).json({msg:"login first"});
          }
 	 	 const updateSlot = await Slot.find({bookedBy:id}).populate({path:'by',select:'username email roles'})
 	     return res.status(200).json(updateSlot);
@@ -124,22 +134,22 @@ export const completeSlot= async (req,res,next)=>{
 	const {completed,rating,by,review} = req.body
 
 	if(id===by){
-		    return res.status(403).json("you can't approve your self");
+		    return res.status(403).json({msg:"you can't approve your self"});
      }
      if(rating===0){
-		    return res.status(403).json("give atleast star rating");	   
+		    return res.status(403).json({msg:"give atleast star rating"});	   
      }
 
 
 	try{
 		 if(!id){
-		 return res.status(200).json("login first");
+		 return res.status(200).json({msg:"login first"});
          }
          if(!ID){
-		 return res.status(200).json("Invalid slot");
+		 return res.status(200).json({msg:"Invalid slot"});
          }
 	 	 const updateSlot = await Slot.findOneAndUpdate({_id:ID},{completed:completed,rating,review},{new:true})
-	     return res.status(200).json(updateSlot);
+	     return res.status(200).json({msg:"Successfully Reviewed"});
 	}catch(err){
 		next(err)
 	}
